requirements.txt
djangorestframework
django-extensions
django-debug-toolbar
unidecode
textract
pandas
requests
psycopg2-binary